Fix amountTrxn conversion to use toString() for accurate string representation

+++ lib/src/models/pay_by_card.dart
@@ -63,7 +63,7 @@ class PayByCardParameters {
     String? tokenCustomerSession,
   }) {
     return PayByCardParameters(
-      amountTrxn: config.amount.toInt().toString(),
+      amountTrxn: config.amount.toString(),
       currencyCodeTrxn: config.currencyCode.toString(),
       merchantId: config.merchantId,
       terminalId: config.terminalId,
